Rename action_cable/*.coffee -> *.js
This extension rename is in preparation for converting these files from
coffeescript to ES2015.

(We are doing the file renames in a separate commit to ensure that the
git history of the files will be preserved - i.e. git will track these
as file renames rather than unrelated file additions/deletions. In
particular, git blame will still trace back to the original authorship.)