Add global configuration for skipping the asset pipeline
This is primarily intended as a tool to speed up tests, since many
tests don't require legitimate asset URLs to be constructed. In those
tests, you could set the global skip_pipeline attribute to true in the
test setup (and reset it in the teardown), to speed up any tests that
render a response but don't test anything that requires the asset
pipeline processing.