Add optional overrides to createConsumer
This will allow callers of createConsumer to pass their own functions
for createConsumer and createWebSocketURL. That is useful because the
default functions have a dependency on the document global, which is
not available inside a Web Worker. So this change moves ActionCable
closer to being able to run from inside a Web Worker.