Only listen for the visibilitychange event if document is defined
This allows ActionCable to be used in a web worker, where the `document`
global is undefined. Without this guard, attempting to use ActionCable
inside a web worker would result in this exception after you try to open
a connection:
```
ReferenceError: document is not defined
```

The visibilitychange listener is mainly a convenience, rather than a
critical piece of the javascript interface, so using ActionCable in a
worker will still work without that listener. (And you could listen for
visibilitychange yourself, then tell the worker to reconnect if you
still want that behavior.)